+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/753>)
 - Add `__repr__` for Dataset
   (<https://github.com/openvinotoolkit/datumaro/pull/750>)
+- Support for exporting as CVAT video format
+  (<https://github.com/openvinotoolkit/datumaro/pull/757>)
 
 ### Changed
 - N/A